hhhh

+++ client/src/main/java/presentation/contentpanel/storepanels/TruckListPanel.java
@@ -94,49 +94,64 @@ public class TruckListPanel extends JPanel implements ActionListener {
      * 载入数据
      */
     protected void refreshList(){
-        try{
-            ArrayList<TruckVO> truckVOs=truckBLService.getTruckList();
-            defaultTableModel.getDataVector().clear();//先清空
-            for (TruckVO vo: truckVOs){
-                String id=vo.getTruckID();
-                String engine=vo.getEngineID();
-                String chepai=vo.getLicenceID();
-                String dipan=vo.getChassisID();
-                String buyTime= Constent.BIRTHDAY_FORMAT.format(vo.getBuyTime());
-                String fuyiTime=vo.getServeTime()+"";
-                String [] data={id,engine,chepai,dipan,buyTime,fuyiTime};
-                defaultTableModel.addRow(data);
+        if (truckBLService!=null) {
+            try{
+                ArrayList<TruckVO> truckVOs=truckBLService.getTruckList();
+                defaultTableModel.getDataVector().clear();//先清空
+                for (TruckVO vo: truckVOs){
+                    String id=vo.getTruckID();
+                    String engine=vo.getEngineID();
+                    String chepai=vo.getLicenceID();
+                    String dipan=vo.getChassisID();
+                    String buyTime= Constent.BIRTHDAY_FORMAT.format(vo.getBuyTime());
+                    Integer fuyiTime=vo.getServeTime();
+                    Object [] data={id,engine,chepai,dipan,buyTime,fuyiTime};
+                    defaultTableModel.addRow(data);
+                }
+                table.revalidate();
+                table.updateUI();
+            } catch (RemoteException e) {
+                new ErrorDialog(parent, "服务器连接超时");
+            } catch (SQLException e) {
+                new ErrorDialog(parent, "SQLException");
             }
-            table.revalidate();
-            table.updateUI();
-        } catch (RemoteException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
+        }
+        else {
+            initBL();
         }
     }
 
     public void actionPerformed(ActionEvent e) {
         if (e.getSource()==addbt){
-            JDialog dialog=new JDialog(parent,"新增车辆信息",true);
-            dialog.getContentPane().add(new TruckInfoPanel(parent, dialog, this, truckBLService));
-            dialog.setLocationRelativeTo(parent);
-            dialog.pack();
-            dialog.setVisible(true);
+            if (truckBLService!=null){
+                JDialog dialog=new JDialog(parent,"新增车辆信息",true);
+                dialog.getContentPane().add(new TruckInfoPanel(parent, dialog, this, truckBLService));
+                dialog.setLocationRelativeTo(parent);
+                dialog.pack();
+                dialog.setVisible(true);
+            }
+            else {
+                initBL();
+            }
         }
         else if (e.getSource()==deletebt){
             int row=table.getSelectedRow();
             if (row==-1){//没有选择任何行
                 new ErrorDialog(parent, "请选择一行待删除条目");
             } else {//选择了待删除的行
-                String id= (String)table.getValueAt(row, 0);
-                try {
-                    truckBLService.deleteTruck(id);
-                    refreshList();
-                } catch (RemoteException e1) {
-                    new ErrorDialog(parent, "网络连接超时");
-                } catch (SQLException e1) {
-                    new ErrorDialog(parent, "SQLException");
+                if (truckBLService!=null){
+                    String id= (String)table.getValueAt(row, 0);
+                    try {
+                        truckBLService.deleteTruck(id);
+                        refreshList();
+                    } catch (RemoteException e1) {
+                        new ErrorDialog(parent, "网络连接超时");
+                    } catch (SQLException e1) {
+                        new ErrorDialog(parent, "SQLException");
+                    }
+                }
+                else{
+                    initBL();
                 }
             }
         }
@@ -145,25 +160,30 @@ public class TruckListPanel extends JPanel implements ActionListener {
             if (row==-1){//没有选择任何行
                 new ErrorDialog(parent, "请选择一行待修改条目");
             } else {//选择了待修改的行
-                String id=(String)table.getValueAt(row, 0);
-                String engine=(String) table.getValueAt(row, 1);
-                String chepai=(String) table.getValueAt(row, 2);
-                String dipan=(String) table.getValueAt(row, 3);
-                Date buyTime=null;
-                try {
-                    buyTime=Constent.BIRTHDAY_FORMAT.parse((String)table.getValueAt(row, 4));
-                } catch (ParseException e1) {
-                    System.out.println("不该发生的情况");
+                if (truckBLService!=null){
+                    String id=(String)table.getValueAt(row, 0);
+                    String engine=(String) table.getValueAt(row, 1);
+                    String chepai=(String) table.getValueAt(row, 2);
+                    String dipan=(String) table.getValueAt(row, 3);
+                    Date buyTime=null;
+                    try {
+                        buyTime=Constent.BIRTHDAY_FORMAT.parse((String)table.getValueAt(row, 4));
+                    } catch (ParseException e1) {
+                        System.out.println("不该发生的情况");
+                    }
+                    int fuyiTime=(Integer) table.getValueAt(row, 5);
+
+                    TruckVO vo=new TruckVO(id,chepai, engine, dipan, buyTime, fuyiTime);
+
+                    JDialog dialog=new JDialog(parent,"修改车辆信息",false);
+                    dialog.getContentPane().add(new TruckModifyPanel(parent,dialog, this, truckBLService, vo));
+                    dialog.setLocationRelativeTo(parent);
+                    dialog.pack();
+                    dialog.setVisible(true);
                 }
-                int fuyiTime=(Integer) table.getValueAt(row, 5);
-
-                TruckVO vo=new TruckVO(id,chepai, engine, dipan, buyTime, fuyiTime);
-
-                JDialog dialog=new JDialog(parent,"修改机构信息",false);
-                dialog.getContentPane().add(new TruckModifyPanel(parent,dialog, this, truckBLService, vo));
-                dialog.setLocationRelativeTo(parent);
-                dialog.pack();
-                dialog.setVisible(true);
+                else {
+                    initBL();
+                }
             }
         }
     }
